feat(auth): wire ?invite= into the signup flow
Splits /auth/signup into a server component (page.tsx) that resolves the invite token via resolveInviteContext and a client component (SignUpClient.tsx) that renders the form. When the URL carries ?invite=<token> and resolution succeeds, the page renders a banner ('X invited <email> to join Y') and pre-fills + locks (disables, not just readOnly) the email field via the new lockedEmail prop on PasskeySignupButton.

Failed resolution still renders signup — signup is independent of invite acceptance per the epic. A stale invite surfaces post-signup as a non-blocking dashboard toast in task 9. The locked email gate prevents the email-mismatch acceptance vector even if a user attempts to override the disabled input via DevTools (server-side acceptance pipe re-validates in task 9).

Task 8 of the drive-invite GDPR + zero-trust epic. Out of scope: signing out a logged-in user with a different email than the invite — handled as a follow-up after the targeted post-login pipe lands (task 10).
