+++ apps/web/src/app/auth/signup/page.tsx
@@ -1,164 +1,29 @@
-"use client";
+import { resolveInviteContext } from "@/lib/auth/invite-resolver";
+import { SignUpClient } from "./SignUpClient";
 
-import { useState, useEffect } from "react";
-import { useRouter } from "next/navigation";
-import Link from "next/link";
-import { motion } from "motion/react";
-import { Mail } from "lucide-react";
-import {
-  AuthShell,
-  AuthDivider,
-  OAuthButtons,
-  GoogleOneTap,
-  PasskeySignupButton,
-  ExternalAuthWaiting,
-} from "@/components/auth";
-import { useAuthCSRF } from "@/hooks/useAuthCSRF";
-import { useOAuthSignIn } from "@/hooks/useOAuthSignIn";
-import { isOnPrem } from "@/lib/deployment-mode";
+interface PageProps {
+  searchParams: Promise<{ invite?: string }>;
+}
 
-export default function SignUp() {
-  // On-prem: redirect to signin (self-registration disabled) - early return to avoid flash
-  if (isOnPrem()) {
-    return <OnPremSignUpRedirect />;
+export default async function SignUp({ searchParams }: PageProps) {
+  const { invite } = await searchParams;
+
+  if (!invite) {
+    return <SignUpClient inviteContext={null} />;
   }
 
-  return <CloudSignUp />;
-}
+  const resolution = await resolveInviteContext({ token: invite, now: new Date() });
+  // Failed resolution still renders signup — the user may have arrived here
+  // outside the consent screen, and signup itself doesn't depend on a valid
+  // invite. Acceptance is independent (handled post-signup); a stale invite
+  // surfaces as a non-blocking dashboard toast (task 9).
+  const inviteContext = resolution.ok
+    ? {
+        driveName: resolution.data.driveName,
+        inviterName: resolution.data.inviterName,
+        email: resolution.data.email,
+      }
+    : null;
 
-function OnPremSignUpRedirect() {
-  const router = useRouter();
-  useEffect(() => {
-    router.replace("/auth/signin?onprem=contact_admin");
-  }, [router]);
-  return null;
-}
-
-function CloudSignUp() {
-  const [error, setError] = useState<string | null>(null);
-  const router = useRouter();
-  const { csrfToken, refreshToken } = useAuthCSRF();
-  const [passkeyLoading, setPasskeyLoading] = useState(false);
-  const {
-    handleGoogleSignIn,
-    handleAppleSignIn,
-    isGoogleLoading,
-    isAppleLoading,
-    isWaitingForExternalAuth,
-    waitingProvider,
-    cancelExternalAuth,
-  } = useOAuthSignIn({
-      onStart: () => setError(null),
-      onError: (msg) => setError(msg),
-    });
-
-  const isAnyLoading = isGoogleLoading || isAppleLoading || passkeyLoading;
-
-  return (
-    <AuthShell>
-      <GoogleOneTap autoSelect={true} cancelOnTapOutside={true} context="signup" />
-
-      {/* Heading */}
-      <motion.div
-        className="mb-8 text-center"
-        initial={{ opacity: 0, y: 16 }}
-        animate={{ opacity: 1, y: 0 }}
-        transition={{ duration: 0.4 }}
-      >
-        <h1 className="text-3xl font-bold tracking-tight text-gray-900 dark:text-white">
-          Get on the same page.
-        </h1>
-        <p className="mt-2 text-sm text-muted-foreground">
-          Create your free workspace in seconds
-        </p>
-      </motion.div>
-
-      {error && (
-        <motion.p
-          className="mb-4 text-center text-sm text-red-500"
-          initial={{ opacity: 0 }}
-          animate={{ opacity: 1 }}
-        >
-          {error}
-        </motion.p>
-      )}
-
-      {/* Passkey signup */}
-      {csrfToken && (
-        <motion.div
-          initial={{ opacity: 0, y: 10 }}
-          animate={{ opacity: 1, y: 0 }}
-          transition={{ delay: 0.2, duration: 0.3 }}
-        >
-          <PasskeySignupButton
-            csrfToken={csrfToken}
-            refreshToken={refreshToken}
-            onEmailExists={() => {
-              router.push("/auth/signin");
-            }}
-            onLoadingChange={setPasskeyLoading}
-            disabled={isAnyLoading}
-          />
-        </motion.div>
-      )}
-
-      <AuthDivider delay={0.3} />
-
-      {/* OAuth buttons */}
-      {isWaitingForExternalAuth ? (
-        <ExternalAuthWaiting provider={waitingProvider} onCancel={cancelExternalAuth} />
-      ) : (
-        <motion.div
-          initial={{ opacity: 0, y: 10 }}
-          animate={{ opacity: 1, y: 0 }}
-          transition={{ delay: 0.35, duration: 0.3 }}
-        >
-          <OAuthButtons
-            onGoogleClick={handleGoogleSignIn}
-            onAppleClick={handleAppleSignIn}
-            disabled={isAnyLoading}
-            isGoogleLoading={isGoogleLoading}
-            isAppleLoading={isAppleLoading}
-          />
-        </motion.div>
-      )}
-
-      {/* Magic link fallback */}
-      <motion.div
-        className="mt-4 text-center"
-        initial={{ opacity: 0 }}
-        animate={{ opacity: 1 }}
-        transition={{ delay: 0.45, duration: 0.3 }}
-      >
-        <Link
-          href="/auth/magic-link"
-          className="inline-flex items-center gap-1.5 text-sm text-muted-foreground transition-colors hover:text-foreground"
-        >
-          <Mail className="h-3.5 w-3.5" />
-          Or sign up with email link
-        </Link>
-      </motion.div>
-
-      {/* Footer */}
-      <motion.div
-        className="mt-8 text-center"
-        initial={{ opacity: 0 }}
-        animate={{ opacity: 1 }}
-        transition={{ delay: 0.55, duration: 0.3 }}
-      >
-        <p className="text-sm text-muted-foreground">
-          Already have an account?{" "}
-          <Link
-            href="/auth/signin"
-            className="font-medium text-blue-600 hover:text-blue-500 dark:text-blue-400 dark:hover:text-blue-300"
-          >
-            Log in
-          </Link>
-        </p>
-      </motion.div>
-
-      {/* ToS acceptance is now an explicit checkbox inside PasskeySignupButton — */}
-      {/* the previous footer text was non-binding and duplicated that gate. */}
-    </AuthShell>
-  );
+  return <SignUpClient inviteContext={inviteContext} />;
 }
